Got gyro turn working in one direction with PD controller
